Refactored code for Locus Report.
