+++ CHEWBBACA/SchemaEvaluator/resources/loci_reports/src/components/AlertMUI.js
@@ -1,0 +1,15 @@
+import Alert from '@mui/material/Alert';
+import Typography from '@mui/material/Typography';
+
+
+const AlertMUI = ({ severity, fontSize, children }) => {
+	return (
+		<Alert severity={severity}>
+			<Typography sx={{ fontSize: fontSize }}>
+				{children}
+			</Typography>
+		</Alert>
+	)
+}
+
+export default AlertMUI;
